Move log file location. Ref #74
